refactor(handler,e2e): split connections, integrations, org files under 300 line limit
